organize modal components
