+++ src/app/[orgSlug]/study/[studyId]/edit-and-resubmit/footer.tsx
@@ -5,8 +5,8 @@ import { useRouter, useParams } from 'next/navigation'
 import { Button, Group } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
 import { CaretLeftIcon } from '@phosphor-icons/react'
-import { AppModal } from '@/components/modal'
-import { SubmitConfirmationModal } from '@/components/submit-confirmation-modal'
+import { AppModal } from '@/components/modals/app-modal'
+import { SubmitConfirmationModal } from '@/components/modals/submit-confirmation-modal'
 import { Routes } from '@/lib/routes'
 import { hasLexicalContent } from '@/lib/lexical'
 import { useEditResubmit } from '@/contexts/edit-resubmit'
