+++ src/app/[orgSlug]/study/[studyId]/proposal/footer.tsx
@@ -4,8 +4,8 @@ import { FC } from 'react'
 import { useRouter, useParams } from 'next/navigation'
 import { Button, Group } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
-import { AppModal } from '@/components/modal'
-import { SubmitConfirmationModal } from '@/components/submit-confirmation-modal'
+import { AppModal } from '@/components/modals/app-modal'
+import { SubmitConfirmationModal } from '@/components/modals/submit-confirmation-modal'
 import { CaretLeftIcon } from '@phosphor-icons/react'
 import { useProposal } from '@/contexts/proposal'
 import { Routes } from '@/lib/routes'
