+++ src/app/[orgSlug]/study/[studyId]/review/proposal-review-view.test.tsx
@@ -13,7 +13,7 @@ import { lexicalJson } from '@/lib/lexical'
 import { memoryRouter } from 'next-router-mock'
 import { useParams } from 'next/navigation'
 import { beforeEach, describe, expect, it } from 'vitest'
-import { ReviewConfirmationModal, REJECTION_WARNING } from '@/components/review-confirmation-modal'
+import { ReviewConfirmationModal, REJECTION_WARNING } from '@/components/modals/review-confirmation-modal'
 import { ProposalReviewView } from './proposal-review-view'
 
 describe('ProposalReviewView', () => {
